+++ Gems/EMotionFX/Code/EMotionFX/Source/DualQuatSkinDeformer.h
@@ -68,7 +68,7 @@ namespace EMotionFX
          * @param node The node where the mesh belongs to during this initialization.
          * @param lodLevel The LOD level of the mesh the mesh deformer works on.
          */
-        void Reinitialize(Actor* actor, Node* node, uint32 lodLevel) override;
+        void Reinitialize(Actor* actor, Node* node, size_t lodLevel) override;
 
         /**
          * Creates an exact clone (copy) of this deformer, and returns a pointer to it.
